moved app into its own folder
