+++ src/app/app.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import React from 'react';
 
-import App from './App';
+import App from './app';
 
 test('renders app header', () => {
   render(<App />);
