+++ src/index.tsx
@@ -2,7 +2,7 @@ import { Global, css } from '@emotion/react';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './App';
+import App from './app';
 import reportWebVitals from './reportWebVitals';
 
 const GlobalCSS = css`
